Corrected some data-types

+++ mp_read_radix.c
@@ -40,9 +40,9 @@ mp_err mp_read_radix(mp_int *a, const char *str, int radix)
 
    /* Try faster version first */
    if (MP_HAS(S_MP_FASTER_READ_RADIX)) {
-      if ((err = s_mp_faster_read_radix(a, str, 0, (int)slen, radix)) != MP_OKAY)                        goto LTM_ERR;
+      if ((err = s_mp_faster_read_radix(a, str, 0, slen, radix)) != MP_OKAY)                             goto LTM_ERR;
    } else if (MP_HAS(S_MP_SLOWER_READ_RADIX)) {
-      if ((err = s_mp_slower_read_radix(a, str, 0, (int)slen, radix)) != MP_OKAY)                        goto LTM_ERR;
+      if ((err = s_mp_slower_read_radix(a, str, 0, slen, radix)) != MP_OKAY)                             goto LTM_ERR;
    }
 
    /* set the sign only if a != 0 */
